Use unsigned int instead of size_t.
git-svn-id: https://subversion.assembla.com/svn/q3/trunk@2197 8af8166b-12a6-a448-a533-9086ace3f9f6

+++ q3/qm/src/model/messagestore.cpp
@@ -204,7 +204,7 @@ bool qm::SingleMessageStore::save(const Message& header,
 		reinterpret_cast<const unsigned char*>(pszBody),
 		SingleMessageStoreImpl::szUnusedSeparator__
 	};
-	size_t nMsgLen[] = {
+	unsigned int nMsgLen[] = {
 		SingleMessageStoreImpl::SEPARATOR_SIZE,
 		nHeaderLen,
 		2,
